# frontend/src/app/core/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecipePresentationIngredientOut": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "type": "number"
          },
          "foodstuff": {
            "$ref": "#/components/schemas/FoodstuffSummaryOut"
          },
          "index": {
            "minimum": 1.0,
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "amount",
          "foodstuff"
        ],
        "title": "RecipePresentationIngredientOut",
        "type": "object"
      },
      "RecipePresentationIngredientResolve": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 9999.0,
            "title": "Amount",
            "type": "number"
          },
          "foodstuffId": {
            "exclusiveMinimum": 0.0,
            "title": "Foodstuffid",
            "type": "integer"
          },
          "index": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "amount",
          "foodstuffId"
        ],
        "title": "RecipePresentationIngredientResolve",
        "type": "object"
      },
      "RecipePresentationOut": {
        "additionalProperties": false,
        "properties": {
          "carbs": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs"
          },
          "fat": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipePresentationIngredientOut"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "kcal": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kcal"
          },
          "preptime": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preptime"
          },
          "protein": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein"
          },
          "servings": {
            "minimum": 1.0,
            "title": "Servings",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RecipePresentationStepOut"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "servings",
          "preptime",
          "kcal",
          "carbs",
          "protein",
          "fat",
          "ingredients",
          "steps"
        ],
        "title": "RecipePresentationOut",
        "type": "object"
      },
      "RecipePresentationResolve": {
        "additionalProperties": false,
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipePresentationIngredientResolve"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "preptime": {
            "anyOf": [
              {
                "maximum": 999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preptime"
          },
          "servings": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Servings",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RecipePresentationStepResolve"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "servings",
          "preptime",
          "ingredients",
          "steps"
        ],
        "title": "RecipePresentationResolve",
        "type": "object"
      },
      "RecipePresentationStepOut": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "index": {
            "minimum": 1.0,
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "description"
        ],
        "title": "RecipePresentationStepOut",
        "type": "object"
      },
      "RecipePresentationStepResolve": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "index": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "description"
        ],
        "title": "RecipePresentationStepResolve",
        "type": "object"
      },
      "RecipeVersionOut": {
        "additionalProperties": false,
        "properties": {
          "carbs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "fat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientOut"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "kcal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kcal"
          },
          "lastModified": {
            "format": "date-time",
            "title": "Lastmodified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "originName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originname"
          },
          "originUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originurl"
          },
          "preptime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preptime"
          },
          "protein": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein"
          },
          "recipeLineageId": {
            "format": "uuid",
            "title": "Recipelineageid",
            "type": "string"
          },
          "recipeVersionId": {
            "format": "uuid",
            "title": "Recipeversionid",
            "type": "string"
          },
          "servings": {
            "title": "Servings",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/RecipeVersionState"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "recipeLineageId",
          "recipeVersionId",
          "state",
          "createdAt",
          "lastModified",
          "name",
          "servings",
          "preptime",
          "originName",
          "originUrl",
          "kcal",
          "carbs",
          "protein",
          "fat",
          "ingredients",
          "steps"
        ],
        "title": "RecipeVersionOut",
        "type": "object"
      },
...
      "RecipeVersionWrite": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientWrite"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "originName": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originname"
          },
          "originUrl": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originurl"
          },
          "preptime": {
            "anyOf": [
              {
                "maximum": 999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preptime"
          },
          "servings": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Servings",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepWrite"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "name",
          "servings"
        ],
        "title": "RecipeVersionWrite",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kochwiki API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/recipe-presentations/resolve": {
      "post": {
        "operationId": "resolve_recipe_presentation_recipe_presentations_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipePresentationResolve"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipePresentationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Recipe Presentation"
      }
    },
    "/recipes": {
      "get": {
        "operationId": "list_recipe_versions_recipes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeVersionOut"
                  },
                  "title": "Response List Recipe Versions Recipes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Recipe Versions"
      },
      "post": {
        "operationId": "create_recipe_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeVersionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Recipe"
      }
    },
    "/recipes/{lineage_id}": {
      "delete": {
        "operationId": "delete_recipe_lineage_recipes__lineage_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Recipe Lineage"
      },
      "get": {
        "operationId": "get_active_recipe_version_recipes__lineage_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Recipe Version"
      }
    },
    "/recipes/{lineage_id}/drafts": {
      "post": {
        "operationId": "create_recipe_draft_recipes__lineage_id__drafts_post",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeVersionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Recipe Draft"
      }
    },
    "/recipes/{lineage_id}/drafts/{version_id}": {
      "delete": {
        "operationId": "discard_recipe_draft_recipes__lineage_id__drafts__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Recipe Draft"
      },
      "put": {
        "operationId": "update_recipe_draft_recipes__lineage_id__drafts__version_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeVersionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Recipe Draft"
      }
    },
    "/recipes/{lineage_id}/drafts/{version_id}/publish": {
      "post": {
        "operationId": "publish_recipe_draft_recipes__lineage_id__drafts__version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Recipe Draft"
      }
    },
    "/recipes/{lineage_id}/history": {
      "get": {
        "operationId": "list_historical_recipe_versions_recipes__lineage_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeVersionOut"
                  },
                  "title": "Response List Historical Recipe Versions Recipes  Lineage Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Historical Recipe Versions"
      }
    },
    "/recipes/{lineage_id}/publish": {
      "post": {
        "operationId": "publish_active_recipe_edit_recipes__lineage_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeVersionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Active Recipe Edit"
      }
    },
    "/recipes/{lineage_id}/versions/{version_id}": {
      "get": {
        "operationId": "get_recipe_version_by_id_recipes__lineage_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lineage_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lineage Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recipe Version By Id"
      }
    },
...
  }
}